+++ src/pages/sign-up/corporate/corporateMember.module.scss
@@ -1,4 +1,4 @@
-@import '../../../components/sign-up/commonSignup.module.scss';
+@import '../../../components/sign-up/commonSignup.module';
 
 .textBlue {
   & > strong {
