+++ src/pages/sign-up/individual/individualMember.module.scss
@@ -1,1 +1,1 @@
-@import '../../../components/sign-up/commonSignup.module.scss';
+@import '../../../components/sign-up/commonSignup.module';
